Unified LIA and LRA solvers
