+++ src/tsolvers/LATHandler.cc
@@ -1,0 +1,31 @@
+//
+// Created by Antti on 21.10.21.
+//
+
+#include "LATHandler.h"
+#include "TreeOps.h"
+#include "LASolver.h"
+#include "OsmtInternalException.h"
+
+LATHandler::LATHandler(SMTConfig & c, ArithLogic & l)
+    : TSolverHandler(c)
+    , logic(l)
+{
+    lasolver = new LASolver(config, logic);
+    SolverId my_id = lasolver->getId();
+    tsolvers[my_id.id] = lasolver;
+    solverSchedule.push(my_id.id);
+}
+
+PTRef LATHandler::getInterpolant(ipartitions_t const & mask, map<PTRef, icolor_t> * labels, PartitionManager & pmanager) {
+    if (logic.isRealLogic()) {
+        return lasolver->getRealInterpolant(mask, labels, pmanager);
+    } else if (logic.isIntegralLogic()) {
+        if (labels == nullptr) {
+            throw OsmtInternalException("LIA interpolation requires partitioning map, but no map was provided");
+        }
+        return lasolver->getIntegerInterpolant(*labels);
+    } else {
+        throw OsmtInternalException("Mixed arithmetic interpolation not supported");
+    }
+}
