+++ src/tsolvers/liasolver/CMakeLists.txt
@@ -1,9 +1,7 @@
 target_sources(tsolvers
-    PUBLIC "${CMAKE_CURRENT_LIST_DIR}/LIASolver.h"
     PUBLIC "${CMAKE_CURRENT_LIST_DIR}/Matrix.h"
     PUBLIC "${CMAKE_CURRENT_LIST_DIR}/LIAInterpolator.h"
 
-    PRIVATE "${CMAKE_CURRENT_LIST_DIR}/LIASolver.cc"
     PRIVATE "${CMAKE_CURRENT_LIST_DIR}/Matrix.cc"
     PRIVATE "${CMAKE_CURRENT_LIST_DIR}/LIAInterpolator.cc"
